Fixed of problem of error in debug mode. version 1.7
git-svn-id: https://plugins.svn.wordpress.org/simple-nivoslider/trunk@950020 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ req/SimpleNivoSliderAdmin.php
@@ -531,11 +531,13 @@ class SimpleNivoSliderAdmin {
 	 */
 	function save_apply_simplenivoslider_postdata( $post_id ) {
 
-		$dataapply = $_POST['simplenivoslider_apply'];
-		if ( $dataapply === 'true' ) {
-			add_post_meta( $post_id, 'simplenivoslider_apply', $dataapply, true );
-		} else if ( $dataapply === ''  || $dataapply === 'false' ) {
-			delete_post_meta( $post_id, 'simplenivoslider_apply' );
+		if ( isset($_POST['simplenivoslider_apply']) ) {
+			$dataapply = $_POST['simplenivoslider_apply'];
+			if ( $dataapply === 'true' ) {
+				add_post_meta( $post_id, 'simplenivoslider_apply', $dataapply, true );
+			} else if ( $dataapply === ''  || $dataapply === 'false' ) {
+				delete_post_meta( $post_id, 'simplenivoslider_apply' );
+			}
 		}
 
 	}
